Refactoring code, by adding custom hooks, removing console logs...
