+++ netlify/functions/reset_pictures/reset_pictures.ts
@@ -31,33 +31,11 @@ export const handler: Handler = async (event, context) => {
       dummyPictures
     )
 
-
-
-    // await pictureModel.collection.drop()
-
-    // const pictures = [...Array(24)].map((_, i) => {
-    //   const date = new Date()
-    //   date.setUTCHours(0, 0, 0, 0)
-    //   date.setUTCMonth(11)
-    //   date.setUTCDate(i + 1)
-    //   return {
-    //     day: i + 1,
-    //     isOpen: false,
-    //     isOpenable: false,
-    //     key: i + 1 + ".jpg",
-    //     date: date,
-    //   } as Picture
-    // })
-
-    // await pictureModel.insertMany(
-    //   pictures
-    // )
     return {
       statusCode: 200,
       body: JSON.stringify(dummyPictures),
     }
   } catch (error) {
-    console.log(error)
     return { statusCode: 500, body: error.toString() }
   }
 }
