+++ netlify/models/models.tsx
@@ -19,7 +19,6 @@ const userSchema = new Schema<DBUser>({
     password: String
 })
 
-
 export const dummyPictureModel = model<Picture>("Picture", pictureSchema, process.env.MONGODB_DUMMY_PICTURES_COLLECTION)
 export const pictureModel = model<Picture>("Picture", pictureSchema, process.env.MONGODB_PICTURES_COLLECTION)
 export const refreshTokenModel = model<Token>("Token", refreshTokenSchema, "refreshTokens")
