+++ src/components/Auth.tsx
@@ -3,8 +3,8 @@ import { FC, useContext } from "react";
 import { GlobalContext } from "../context";
 
 export const Auth: FC = () => {
-    const context = useContext(GlobalContext)
-    if (!context.authorized) {
+    const { authorized } = useContext(GlobalContext)
+    if (!authorized) {
         return <Navigate to="/login" />;
     }
     return <Outlet />;
