+++ src/components/Calendar/Calendar.tsx
@@ -1,13 +1,9 @@
-import { FC, useEffect } from "react"
+import { FC } from "react"
 
 import CalendarComponent from "./CalendarComponent";
 
 export const Calendar: FC = () => {
     const calendarTest: boolean = false;
-    useEffect(() => {
-        console.log("Calendar update")
-        console.log(calendarTest)
-    })
 
     return <CalendarComponent test={calendarTest} />
 }
