+++ src/components/Calendar/FullscreenDialog.tsx
@@ -1,10 +1,5 @@
-import * as React from 'react';
-import Button from '@mui/material/Button';
 import Dialog from '@mui/material/Dialog';
-import ListItemText from '@mui/material/ListItemText';
-import ListItem from '@mui/material/ListItem';
 import List from '@mui/material/List';
-import Divider from '@mui/material/Divider';
 import AppBar from '@mui/material/AppBar';
 import Toolbar from '@mui/material/Toolbar';
 import IconButton from '@mui/material/IconButton';
@@ -12,15 +7,16 @@ import Typography from '@mui/material/Typography';
 import CloseIcon from '@mui/icons-material/Close';
 import Slide from '@mui/material/Slide';
 import { TransitionProps } from '@mui/material/transitions';
-import { Dispatch, FC, SetStateAction, forwardRef, useState } from 'react';
-import { Picture } from '../types/types';
-import { CDN_URL } from '../constants';
+import { Dispatch, FC, Ref, SetStateAction, forwardRef } from 'react';
+import { Picture } from '../../types/types';
+import { CDN_URL } from '../../constants';
+import { usePicture } from '../../hooks/usePicture';
 
 const Transition = forwardRef(function Transition(
     props: TransitionProps & {
         children: React.ReactElement;
     },
-    ref: React.Ref<unknown>,
+    ref: Ref<unknown>,
 ) {
     return <Slide direction="up" ref={ref} {...props} />;
 });
@@ -32,18 +28,14 @@ type DialogProps = {
 }
 
 export const FullScreenDialog: FC<DialogProps> = ({ picture, open, setOpen }) => {
-
+    const { imageSRC } = usePicture({ picture })
 
     const handleClose = () => {
-        console.log("CLOSE. IsOPEN:", open)
         setOpen((prev) => !prev);
     };
 
-    const imageSRC = (picture.key.split(".jpg").length == 2) ? CDN_URL + picture.key : picture.key
-
-
     return (
-        <React.Fragment>
+        <>
             <Dialog
                 fullScreen
                 open={open}
@@ -69,7 +61,7 @@ export const FullScreenDialog: FC<DialogProps> = ({ picture, open, setOpen }) =>
                     <img src={imageSRC} width={"100%"} />
                 </List>
             </Dialog>
-        </React.Fragment>
+        </>
     );
 }
 export default FullScreenDialog
