+++ src/components/Home.tsx
@@ -1,11 +1,10 @@
-import { FC, useContext, useEffect, useState } from "react";
+import { FC, useContext } from "react";
 import { CountdownTimer } from "./CountdownTimer/CountdownTimer";
 import { Box } from "@mui/material";
 import { GlobalContext } from "../context";
 
 export const Home: FC = () => {
     const { startingDate } = useContext(GlobalContext)
-
 
     return <Box
         style={{
