+++ src/utils/utils.tsx
@@ -2,22 +2,18 @@ import { cloneDeep } from "lodash"
 
 export const shuffle = (array: any[], seed: number) => {
     var copy = cloneDeep(array)
-    // <-- ADDED ARGUMENT
     var m = copy.length, t, i;
 
     // While there remain elements to shuffle…
     while (m) {
-
         // Pick a remaining element…
-        i = Math.floor(random(seed) * m--);        // <-- MODIFIED LINE
-
+        i = Math.floor(random(seed) * m--);
         // And swap it with the current element.
         t = copy[m];
         copy[m] = copy[i];
         copy[i] = t;
-        ++seed                                     // <-- ADDED LINE
+        ++seed
     }
-
     return copy;
 }
